Added a couple more tests
